requirements.txt
Django==5.2.6
djangorestframework==3.16.1
python3-saml==1.16.0
django-allauth==65.4.1
psycopg[binary,pool]==3.2.4
uwsgi==2.0.28
django-redis==5.4.0
celery==5.4.0
drf-yasg==1.21.8
Pillow==11.1.0
django-imagekit==5.0.0
markdown==3.7
django-filter==24.3
filetype==1.2.0
django-mptt==0.16.0
crispy-bootstrap5==2024.10
requests==2.32.3
PyJWT[crypto]==2.10.1
django-celery-email==3.0.0
m3u8==6.0.0
django-debug-toolbar==5.0.1
pre-commit==4.1.0
django-jazzmin==3.0.1
pysubs2==1.8.0
sentry-sdk[django]==2.23.1
django-tinymce==4.1.0
